Add decimal input to CHU offset

+++ src/databits.h
@@ -21,7 +21,7 @@
 
 extern bool chu_do_systime;
 extern bool chu_exit_on_sync;
-extern int chu_seconds_offset;
+extern float chu_seconds_offset;
 
 // Reverses the ordering of the bits on an integer
 static inline unsigned long long
